+++ requirements.txt
@@ -28,3 +28,4 @@ uWSGI==2.0.21
 Werkzeug==2.2.3
 xxhash==3.2.0
 zipp==3.8.1
+requests~=2.27.1
