--- setup.sh
@@ -1,3 +1,0 @@
-git clone https://github.com/kazet/wpgarlic.git
-
-cp print_findings.py wpgarlic/
